demos/data_loader/requirements.txt
llama_index==0.10.6
llama-index-embeddings-huggingface==0.1.1
llama-index-vector-stores-chroma==0.1.1
tree-sitter
tree-sitter-languages
psycopg2
click
requests
python-dotenv
